+++ source/solution/application/model/glacier_transfer_meta_model.py
@@ -46,3 +46,6 @@ class GlacierTransferMetadata(GlacierTransferMetadataRead):
     s3_destination_key: str | None = Model.field(
         ["s3_destination_key", "S"], optional=True
     )
+    vault_account_id: str | None = Model.field(
+        ["vault_account_id", "S"], optional=True
+    )